requirements.txt
google-genai>=1.0.0
google-cloud-aiplatform>=1.38.0
edge-tts>=6.1.0
Pillow>=10.0.0
python-dotenv>=1.0.0
